vhs-platform/requirements-base.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.0
beautifulsoup4==4.11.1
blessed==1.19.1
bottle==0.12.23
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.4
cssselect2==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
Django==4.0.4
django-admin-extra-buttons==1.5.4
django-admin-searchable-dropdown==1.2
django-autocomplete-light==3.9.1
django-classy-tags==3.0.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-native-tags==0.5.3
django-nested-admin==3.4.0
django-picklefield==3.1
django-polymorphic==3.1.0
fontawesomefree==6.1.1
frozendict==2.3.4
idna==3.3
iiif-prezi==0.3.0
jinxed==1.2.0
lxml==4.8.0
packaging==21.3
pdf2image==1.16.2
pikepdf==6.2.0
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyLD==2.0.3
pyparsing==3.0.9
PyPDF2==2.10.4
python-dateutil==2.8.2
python-environ==0.4.54
python-magic==0.4.27
python-monkey-business==1.0.0
redis==3.5.3
reportlab==3.6.9
requests==2.28.1
setuptools==67.1.0
six==1.16.0
sorl-thumbnail==12.8.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
svglib==1.2.1
tinycss2==1.1.1
tornado==6.1
tripoli==2.0.0
tzdata==2022.1
Unidecode==1.1.2
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
